Renaming packages from github.com/cab105 to github.com/exoscale
