+++ src/github.com/exoscale/terraform-provider-exoscale/exoscale/affinity_resource.go
