+++ src/github.com/exoscale/terraform-provider-exoscale/main.go
@@ -3,7 +3,7 @@ package main
 /* Bootstrap the plugin for Terraform */
 
 import (
-	"github.com/cab105/terraform-provider-exoscale/exoscale"
+	"github.com/exoscale/terraform-provider-exoscale/exoscale"
 	"github.com/hashicorp/terraform/plugin"
 )
